refactor(Game.java): Game 객체에서 책임져야 하는 메소드들을 GameController에서 가져오도록 수정
- 경주하는 자동차 리스트 생성 메소드 : Game에서 담당
- 경주 자동차 이동 및 멈춤 메소드 : Game에서 담당
- 자동차 리스트 생성 테스트 : GameControllerTest 코드에 있던 테스트 코드를 GameTest로 수정
- 경주의 우승자 구하는 메소드 : Game에서 담당
